Change the dependency on better-sqlite3 dynamic
Before this change, pe-sqlite-for-rxdb requires better-sqlite3
statically, which prevents it from being used in contexts where
better-sqlite3 cannot be used.  Specifically, Svelte tries to load
better-sqlite3 and fails because of the static dependency.  Thus, this
change allows Svelte to use pe-sqlite-for-rxdb (if the developer is
careful.)

The additional changes, e.g., to vite.config.ts, solve errors that
appeared in the process of making this depency dynamic.