Make vitest and eslint/tsc happy
This change exists to make tools happy.

eslint/tsc was unhappy because the `schemaTyped` variable was only used
to hold a type, thus unused in JavaScript.  The use is valid, so we will
ignore the warning.

vitest was unhappy because a newer version of vitest will not
automatically await expect().resolves calls.  This is probably good
because using explicit awaits will allow me to address any livelocks
that may occur.